+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # Change log
 
-## [0.4.2] - 2018-5-17
+## [0.4.6] - 2018-5-30
 
-- Add firestore query feature
+- Improve document exists method
 
 ## [0.4.4] - 2018-5-25
 
 - Restructure firestore functions
+
+## [0.4.2] - 2018-5-17
+
+- Add firestore query feature
+
+